update controllers, test with memory fix
