--- src/main/controllers/reasonable_adjustments/ReasonableAdjustmentsController.ts
@@ -1,31 +1,0 @@
-import { Response } from 'express';
-
-import { Form } from '../../components/form/form';
-import { AppRequest } from '../../definitions/appRequest';
-import { PageUrls, TranslationKeys } from '../../definitions/constants';
-import { FormContent, FormFields } from '../../definitions/form';
-import { assignFormData, getPageContent, handleSessionErrors, setUserCase } from '../helpers';
-
-export default class ReasonableAdjustmentsController {
-  private readonly form: Form;
-
-  constructor(private readonly reasonableAdjustmentsContent: FormContent) {
-    this.form = new Form(<FormFields>this.reasonableAdjustmentsContent.fields);
-  }
-
-  public post = (req: AppRequest, res: Response): void => {
-    setUserCase(req, this.form);
-    handleSessionErrors(req, res, this.form, PageUrls.DOCUMENTS);
-  };
-
-  public get = (req: AppRequest, res: Response): void => {
-    const content = getPageContent(req, this.reasonableAdjustmentsContent, [
-      TranslationKeys.COMMON,
-      TranslationKeys.REASONABLE_ADJUSTMENTS,
-    ]);
-    assignFormData(req.session.userCase, this.form.getFormFields());
-    res.render('generic-form-template', {
-      ...content,
-    });
-  };
-}
