+++ src/test/unit/controller/SupportController.test.ts
@@ -1,6 +1,4 @@
-import { atLeastOneFieldIsChecked } from '../../../main/components/form/validator';
-import SupportController from '../../../main/controllers/support/SupportController';
-import { FormContent } from '../../../main/definitions/form';
+import SupportController from '../../../main/controllers/SupportController';
 import { mockRequest } from '../mocks/mockRequest';
 import { mockResponse } from '../mocks/mockResponse';
 
@@ -10,26 +8,8 @@ describe('Support Controller', () => {
     common: {},
   };
 
-  const mockFormContent = {
-    fields: {
-      support: {
-        id: 'support',
-        type: 'checkboxes',
-        labelHidden: true,
-        validator: jest.fn(atLeastOneFieldIsChecked),
-        values: [
-          {
-            id: 'support',
-            name: 'support',
-            value: 'anything',
-          },
-        ],
-      },
-    },
-  } as unknown as FormContent;
-
   it('should render the "I need support" page', () => {
-    const controller = new SupportController(mockFormContent);
+    const controller = new SupportController();
 
     const response = mockResponse();
     const request = mockRequest({ t });
@@ -44,7 +24,7 @@ describe('Support Controller', () => {
       const errors = [{ propertyName: 'support', errorType: 'required' }];
       const body = { support: [''] };
 
-      const controller = new SupportController(mockFormContent);
+      const controller = new SupportController();
 
       const req = mockRequest({ body });
       const res = mockResponse();
