+++ src/test/unit/controller/TravelController.test.ts
@@ -1,6 +1,4 @@
-import { atLeastOneFieldIsChecked } from '../../../main/components/form/validator';
-import TravelController from '../../../main/controllers/travel/TravelController';
-import { FormContent } from '../../../main/definitions/form';
+import TravelController from '../../../main/controllers/TravelController';
 import { mockRequest } from '../mocks/mockRequest';
 import { mockResponse } from '../mocks/mockResponse';
 
@@ -10,26 +8,8 @@ describe('Travel Controller', () => {
     common: {},
   };
 
-  const mockFormContent = {
-    fields: {
-      travel: {
-        id: 'travel',
-        type: 'checkboxes',
-        labelHidden: true,
-        validator: jest.fn(atLeastOneFieldIsChecked),
-        values: [
-          {
-            id: 'travel',
-            name: 'travel',
-            value: 'anything',
-          },
-        ],
-      },
-    },
-  } as unknown as FormContent;
-
   it('should render the "I need help travelling" page', () => {
-    const controller = new TravelController(mockFormContent);
+    const controller = new TravelController();
 
     const response = mockResponse();
     const request = mockRequest({ t });
@@ -44,7 +24,7 @@ describe('Travel Controller', () => {
       const errors = [{ propertyName: 'travel', errorType: 'required' }];
       const body = { travel: [''] };
 
-      const controller = new TravelController(mockFormContent);
+      const controller = new TravelController();
 
       const req = mockRequest({ body });
       const res = mockResponse();
